Calculator.Classes/Calculator.cs
namespace E01.Classes
{
    public class Calculator
    {
        public double Add(double a, double b)
        {
            return a + b;
        }
        public double Subtract(double a, double b)
        {
            return a - b;
        }
        public double Multiply(double a, double b)
        {
            return a * b;
        }
        public double Power(double x, double exp)
        {
            return Math.Pow(x,exp);
        }
        public double Division(double x, double y)
        {
            if (y == 0)
            {
                throw new DivideByZeroException();
            }
            return x / y;
        }
        public double Squareroot(double x)
        {
            return Math.Sqrt(x);
        }
    }
}